fix(build): fixed packaging and package.json fields
