+++ package.json
@@ -1,21 +1,39 @@
 {
   "name": "ngxs",
   "version": "2.0.0-rc.20",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ngxs/store.git"
+  },
+  "keywords": ["ngxs", "redux", "state", "rxjs", "angular", "ngx", "angular2", "cqrs", "store", "state-mangement", "event-stream"],
+  "author": "Austin McDaniel",
+  "contributors": [
+    { "name": "Danny Blue" },
+    { "name": "Leon Radley" }
+  ],
   "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/ngxs/store/issues"
+  },
+  "homepage": "https://github.com/ngxs/store#readme",
   "packageScope": "@ngxs",
   "packages": ["packages/store", "packages/logger-plugin", "packages/devtools-plugin", "packages/storage-plugin"],
   "scripts": {
     "// - APPS": "Run Apps in dev and with packaged modules",
-    "serve:dev": "ng serve --aot",
+    "start": "ng serve --aot",
     "serve:integration": "ng serve --prod --app 1 #requires yarn package",
 
     "// - BUILDING": "Run Apps in dev and with packaged modules",
-    "package": "ts-node tools/package && ts-node tools/set-version",
+    "package": "ts-node tools/package && ts-node tools/set-metadata",
     "build:integration": "ng build --prod --app 1 #requires yarn package",
 
     "// - TESTING": "Test all modules",
     "test": "ng test",
     "test:integration": "ng test --app 1 #requires yarn package",
+
+    "// - CI": "Ci Testing",
+    "test:ci": "ng test --code-coverage --progress=false --browsers ChromeHeadless --single-run",
+    "test:ci:integration": "ng test --progress=false --preserve-symlinks --app 1 --single-run && ng build --prod --progress=false --preserve-symlinks --app 1",
 
     "// - UTILS": "Utility Scripts",
     "lint": "ng lint"
