+++ packages/devtools-plugin/package.json
@@ -1,8 +1,7 @@
 {
   "name": "@ngxs/devtools-plugin",
+  "description": "redux devtools plugin for @ngxs/store",
   "version": "0.0.0",
-  "license": "MIT",
-  "private": true,
   "peerDependencies": {
     "@ngxs/store": "^0.0.0",
     "@angular/core": "^5.0.0"
